% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apim-package.R
\docType{package}
\name{apim-package}
\alias{apim}
\alias{apim-package}
\title{apim: Actor-Partner Interdependence Model Convenience Functions}
\description{
Utilities for Actor-Partner Interdependence Models (APIM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RobAckerman/apim}
  \item Report bugs at \url{https://github.com/RobAckerman/apim/issues}
}

}
\author{
\strong{Maintainer}: Robert Ackerman \email{Robert.Ackerman@utdallas.edu}

Authors:
\itemize{
  \item Robert Ackerman \email{Robert.Ackerman@utdallas.edu}
}

}
\keyword{internal}
